Revert the license change, keeping the code under the MIT License.
The intent of changing the license to AGPL was to promote an open
community of indie MMORPGs, ensuring that no closed-source fork
would come to dominate. After reflecting on this decision for a
few months, however, I realize that I value truly free and open
software more than the legal ability to enforce openness. While I
see the value in the AGPL for e.g. technologies that might be
adopted by a cloud provider, I no longer see a benefit - and
potentially even see harm - in adopting such a license for what is
a hobbyist game engine.1