Partial debugging of light map fragment shader.
Beware of this shader. It's neutralized right now
but will cause a GPU fault if the later elements
in the uniform block are accessed. I believe this
is a bug in Veldrid with how dynamic uniform
buffers are handled in the Vulkan backend, but I
am still investigating. I will make a patch to
fix this in Veldrid if that turns out to be the
case.